Implement Renaming commands system
This may be a unstabled build, so don’t expect this build to be bug free.
This is just an experiment to see if this could be possible or not. As far as we know, there’s no plugins out there that support Customising your own commands, or even Faction commands. (We could be wrong.)