Fixed Factions chat issue - Hover to see
There was a big which was in this plugin a while ago, where if you were in the faction chat, and you'd /f leave, or /f del during the faction chat process, it will keep you in the faction chat, but will not let you in the public chat. This was due to an issue occurring with Faction chat not being removed from the player upon deletion / kicked / leave or even force delete events. The developers probably forgot to implement, and fix this a while back. Some people were to complain about this issue, even now. So this issue should now be fixed. If any more issues occur, please open a new issue. Thank you.