Beta: Added detections onLoad(): & onEnable():
Also, implemented getInstance(): which should allow more-use of the plugin’s API.
I do plan on making this a more-freedom plugin very soon, where those who want to use the plugin API for creating their own factions plugin, then they can. But that’s something we plan in the future.
Along with that, I am using this branch to beta test new exclusive features to the plugin. Once I’ve completed this update, it’ll be transferred to the new-methods branch.