Commands: Fixed many bugs and improvements
Added new function getLeader() as a new check for the /f promote bug.
Added some new checks for isLeader($sender->getName() == true){
Added a new check for isOfficer($sender->getName() == true){ just in case if the bug ever comes back.
Renamed getPlayer() to getPlayerExact() for the $promoted, $promotedName, and $demotedName. That way, it’ll detect the actual username instead of just the getPlayer() which only shows the incorrect username.
Added }elseif( checks to the /f promote command.
Added $leaderName variable.
Fixed many bugs and improvements.
WARNING: This build may be unstabled. If it comes up as a build error, please wait for a OK lint so you can download it without any issues upon enabling, and syntax errors.
A new build for the syntax error will be released soon (If any syntax errors.)