Added new checks: HOVER TO SEE
We've added the following checks to the plugin: checkConfigurations(), checkPlugins(), and registerEvents(), all in onEnable() function. Removed onLoad() function to fix a bug where the plugin would not enable due to the listener not being able to register while not enabled.
I've also added all the checks in onEnable(), so I've done:
protected function onEnable() : void {
$this->registerEvents();
$this->checkConfigurations(); 
$this->checkPlugins();
So that way, it registers all the functions to onEnable() state.