Why was this here?
Fix a weird duplication command code. I've never seen this there before.
Those command usage were both exactly the same, and somehow has been added to this plugin. No wonder why people had issues with this plugin.
This build "should" fix: #55, fix #53, and fix #52 
This build is also currently untested, so please test this to see if it fixes or not.
Thank you, and sorry for the long wait!