Added tick option for updateTagTask
By default, this is setted to 20, which is 1 second. This may be complicated if you don’t know how ticks work. It’s recommended to set the tick option to 20, but it is your choice. This option allows you to edit how many ticks in seconds it’ll update the faction tag.