FACTIONSPRO v2.1.4 UPDATE — hover to see * Remove UpdateNotifier by JackMD (May add this back in the future) * Add new function: checkUpdate() * The plugin now notifies console if a new update from this plugin is available. * Added PluginDescription use namespace. * Added new config option: update-checker. Either set this to true or false. * Fixed some concerns regarding plugin loading issues.
This update ensures that you won’t miss a new update within this plugin. If a new update is available, it should notify console regarding this change. You may choose whether to install a new update of this plugin.
In the near future, I plan on adding more releases to github release tab real soon, and you can get notified within console to go to the releases tab for more stable release of this plugin. This build has not been tested yet. Thank you!