CancelTask for that specific player..
if a player is not in a faction. This means if a player isn't in a faction, then the task will be cancelled for that specific player (Untested, and may change).
This is to prevent useless empty faction names being used by this task if there's no faction. This is currently untested, so please test it out on a new build, and try again.